requirements.txt
pyyaml
pytest
pylint
mock